# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Studio EMAR API
  version: 0.1.0
  description: >
    Contrato da FASE 5 (auth, students, schedules, bookings, credits).
    Sem dashboard e sem join na lista de espera.
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Serviço no ar
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /auth/login:
    post:
      summary: Login com e-mail e senha
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sessão JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          description: Credenciais inválidas ou primeiro acesso pendente

  /auth/first-access:
    post:
      summary: Definir senha no primeiro acesso
      operationId: firstAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirstAccessRequest'
      responses:
        '200':
          description: Senha definida e sessão JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '409':
          description: Conta já possui senha

  /auth/refresh:
    post:
      summary: Renovar o access token
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nova sessão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          description: Refresh inválido

  /auth/recover:
    post:
      summary: Pedir recuperação de senha
      operationId: recoverPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverRequest'
      responses:
        '200':
          description: Sempre ok (não revela se o e-mail existe)

  /auth/reset-password:
    post:
      summary: Definir senha com token de recuperação
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Senha redefinida e sessão JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          description: Token inválido ou expirado

  /me:
    get:
      summary: Usuário autenticado
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuário sem passwordHash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /plans:
    get:
      summary: Planos
      operationId: listPlans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Planos para cadastro de aluno

  /students:
    get:
      summary: Listar alunos
      operationId: listStudents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alunos
    post:
      summary: Criar conta de aluno
      operationId: createStudent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Aluno criado com mustSetPassword
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: E-mail já cadastrado

  /students/{id}:
    get:
      summary: Detalhe do aluno
      operationId: getStudent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Aluno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /students/{id}/bookings:
    get:
      summary: Reservas do aluno (treinador)
      operationId: listStudentBookings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Reservas

  /students/{id}/credits:
    get:
      summary: Créditos do aluno (treinador)
      operationId: listStudentCredits
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Créditos

  /time-slots/{id}:
    get:
      summary: Detalhe do horário
      operationId: getTimeSlot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Horário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'

  /time-slots/{id}/bookings:
    get:
      summary: Participantes do horário (treinador)
      operationId: listTimeSlotBookings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Reservas com aluno. O aluno não usa este path.

  /time-slots/{id}/waitlist:
    get:
      summary: Lista de espera FIFO (treinador)
      operationId: listWaitlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Fila. Sem join nesta fase.

  /recurring-slots:
    get:
      summary: Agenda recorrente
      operationId: listRecurringSlots
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Horários do plano
    post:
      summary: Incluir horário recorrente
      operationId: addRecurringSlot
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Horário criado
        '409':
          description: Já existe

  /recurring-slots/{id}:
    delete:
      summary: Remover horário recorrente
      operationId: removeRecurringSlot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Removido

  /closures:
    get:
      summary: Fechamentos
      operationId: listClosures
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fechamentos
    post:
      summary: Informar fechamento
      operationId: createClosure
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Fechamento. grantsCredit opcional (RN-019).

  /credits:
    get:
      summary: Todos os créditos (treinador)
      operationId: listCredits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Créditos

  /credits/{id}/annulments:
    post:
      summary: Anular crédito
      operationId: annulCredit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Crédito anulado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
        '409':
          description: Crédito não pode ser anulado

  /me/bookings:
    get:
      summary: Agenda do aluno autenticado
      operationId: listMyBookings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de reservas do aluno
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'

  /bookings/{id}/cancellations:
    post:
      summary: Cancelar uma reserva
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Cancelamento registrado. generatedCredit depende de RN-012.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cancellation'
        '409':
          description: Reserva não pode ser cancelada

  /me/credits:
    get:
      summary: Créditos do aluno autenticado
      operationId: listMyCredits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Créditos com rastreio (RN-010)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credit'

  /credits/{id}/redemptions:
    post:
      summary: Usar um crédito em um horário disponível
      operationId: redeemCredit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timeSlotId]
              properties:
                timeSlotId:
                  type: string
      responses:
        '201':
          description: Reserva MAKEUP criada e crédito consumido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '409':
          description: Sem vaga, crédito inválido ou expirado

  /time-slots:
    get:
      summary: Horários e vagas
      description: O aluno vê ocupação numérica, não os nomes dos demais alunos.
      operationId: listTimeSlots
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Horários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    FirstAccessRequest:
      type: object
      required: [email, password, confirmPassword]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 6

    RecoverRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      required: [token, password, confirmPassword]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 6

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    AuthSession:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    CreateStudentRequest:
      type: object
      required: [name, email, planId]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        planId:
          type: string

    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [STUDENT, TRAINER, ADMIN]
        planId:
          type: string
        mustSetPassword:
          type: boolean

    Booking:
      type: object
      required: [id, studentId, timeSlotId, kind, status]
      properties:
        id:
          type: string
        studentId:
          type: string
        timeSlotId:
          type: string
        kind:
          type: string
          enum: [REGULAR, MAKEUP]
        status:
          type: string
          enum: [CONFIRMED, CANCELLED, NO_SHOW]

    Cancellation:
      type: object
      required: [id, bookingId, cancelledAt, cancelledBy, generatedCredit]
      properties:
        id:
          type: string
        bookingId:
          type: string
        cancelledAt:
          type: string
          format: date-time
        cancelledBy:
          type: string
          enum: [STUDENT, TRAINER]
        generatedCredit:
          type: boolean
        creditId:
          type: string

    Credit:
      type: object
      required:
        [id, studentId, source, generatedAt, expiresAt, status]
      properties:
        id:
          type: string
        studentId:
          type: string
        source:
          type: string
          enum: [CANCELLATION, TRAINER_CANCELLATION, CLOSURE_COMPENSATION]
        generatedAt:
          type: string
          format: date-time
        originBookingId:
          type: string
        originClosureId:
          type: string
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [AVAILABLE, USED, EXPIRED, ANNULLED]
        usedAt:
          type: string
          format: date-time
        usedBookingId:
          type: string
        annulledAt:
          type: string
          format: date-time
        annulledByUserId:
          type: string

    TimeSlot:
      type: object
      required:
        [id, startsAt, endsAt, capacity, enrolledCount, status, classType, trainerId]
      properties:
        id:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        capacity:
          type: integer
        enrolledCount:
          type: integer
        status:
          type: string
          enum: [OPEN, FULL, CLOSED]
        classType:
          type: string
        trainerId:
          type: string

    StudioClosure:
      type: object
      required: [id, startsOn, endsOn, reason, createdByUserId, grantsCredit]
      properties:
        id:
          type: string
        startsOn:
          type: string
          format: date
        endsOn:
          type: string
          format: date
        reason:
          type: string
        createdByUserId:
          type: string
        grantsCredit:
          type: boolean
          default: false

    WaitlistEntry:
      type: object
      required: [id, timeSlotId, studentId, position, enqueuedAt, status]
      properties:
        id:
          type: string
        timeSlotId:
          type: string
        studentId:
          type: string
        position:
          type: integer
        enqueuedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [WAITING, PROMOTED, CANCELLED]

    OccupancyDashboard:
      type: object
      required: [metrics, byHour, byWeekday]
      properties:
        metrics:
          type: object
          required:
            [studentsToday, occupancyPercent, freeSpots, cancellations, makeups]
          properties:
            studentsToday:
              type: integer
            occupancyPercent:
              type: number
            freeSpots:
              type: integer
            cancellations:
              type: integer
            makeups:
              type: integer
        byHour:
          type: array
          items:
            type: object
            required: [hour, occupancyPercent]
            properties:
              hour:
                type: string
              occupancyPercent:
                type: number
        byWeekday:
          type: array
          items:
            type: object
            required: [weekday, occupancyPercent]
            properties:
              weekday:
                type: string
              occupancyPercent:
                type: number